Lesson4 Task A,B,C;
