+++ src/by/it/ivanstets/ivanstetsles2/TaskA2.java
@@ -1,0 +1,22 @@
+package by.it.ivanstets.ivanstetsles2;
+
+/*
+напишите тело main программы которая выводит на экран 5 строк
+"Я начинаю изучать Java!"
+
+Для ручной проверки запустите программу Ctrl+Shift+F10
+Для автоматической проверки откройте и запустите класс Testing
+*/
+
+class TaskA2 {
+    public static void main(String[] args) {
+        for (int i = 0; i <5; i++) {
+            System.out.println("Я начинаю изучать Java!");
+        }
+
+
+
+
+
+    }
+}
