+++ src/by/it/ivanstets/ivanstetsles2/TaskB1.java
@@ -1,0 +1,37 @@
+package by.it.ivanstets.ivanstetsles2;
+
+/*
+Выводим квадрат числа
+Напишите программу, которая считывает с клавиатуры целое число a и выводит квадрат этого числа (a * a).
+Для считывания данных с клавиатуры используйте метод nextInt() объекта класса Scanner.
+
+Создать Scanner можно так:
+Scanner sc=new Scanner(System.in);
+
+Прочитать число можно так:
+int i=sc.nextInt();
+
+Требования:
+1. Программа должна выводить текст.
+2. В программе необходимо создать объект типа Scanner.
+3. Программа должна считывать число типа Int с клавиатуры.
+4. Программа должна выводить квадрат считанного числа.
+
+ */
+
+
+import java.util.Scanner;
+
+class TaskB1 {
+    public static void main(String[] args) {
+        Scanner HP = new Scanner(System.in);
+        int i = HP.nextInt();
+        System.out.println(i*i);
+
+    }
+
+
+
+
+
+}
