+++ src/by/it/ivanstets/ivanstetsles2/Testing02.java
@@ -1,0 +1,251 @@
+package by.it.ivanstets.ivanstetsles2;
+
+import org.junit.Test;
+
+import java.io.*;
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
+
+import static org.junit.Assert.*;
+
+@SuppressWarnings("all")
+
+//поставьте курсор на следующую строку и нажмите Ctrl+Shift+F10
+public class Testing02 {
+
+
+    @Test(timeout = 1500)
+    public void testTaskA1() throws Exception {
+        run("").include("Hello world!");
+    }
+
+    @Test(timeout = 1500)
+    public void testTaskA2() throws Exception {
+        run("").include(
+                "Я начинаю изучать Java!\n" +
+                        "Я начинаю изучать Java!\n" +
+                        "Я начинаю изучать Java!\n" +
+                        "Я начинаю изучать Java!\n" +
+                        "Я начинаю изучать Java!\n"
+        );
+    }
+
+    @Test(timeout = 1500)
+    public void testTaskA3() throws Exception {
+        run("").include("3*3+4*4=25");
+    }
+
+    @Test(timeout = 1500)
+    public void testTaskB1() throws Exception {
+        run("7").include("49");
+    }
+
+    @Test(timeout = 1500)
+    public void testTaskB2() throws Exception {
+        run("").include("20");
+    }
+
+    @Test(timeout = 1500)
+    public void testTaskB3() throws Exception {
+        run("").include("C Новым Годом");
+    }
+
+    @Test(timeout = 1500)
+    public void testTaskC1() throws Exception {
+        run("7\n3\n").include("Sum = 10\n");
+    }
+
+    @Test(timeout = 1500)
+    public void testTaskC2() throws Exception {
+        run("34\n26\n").include(
+                "DEC:34+26=60\n" +
+                        "BIN:100010+11010=111100\n" +
+                        "HEX:22+1a=3c\n" +
+                        "OCT:42+32=74\n");
+    }
+
+    @Test(timeout = 1500)
+
+    public void testTaskC3() throws Exception {
+        run("75\n").include("29.51\n");
+        Testing02 t = run("100\n").include("39.35\n");
+        Method m = checkMethod(t.aClass.getSimpleName(), "getWeight", int.class);
+        assertEquals((Double) m.invoke(null, 100), 39.35, 1e-100);
+        assertEquals((Double) m.invoke(null, 75), 29.51, 1e-100);
+    }
+
+    /*
+===========================================================================================================
+НИЖЕ ВСПОМОГАТЕЛЬНЫЙ КОД ТЕСТОВ. НЕ МЕНЯЙТЕ В ЭТОМ ФАЙЛЕ НИЧЕГО.
+Но изучить как он работает - можно, это всегда будет полезно.
+===========================================================================================================
+ */
+    //-------------------------------  методы ----------------------------------------------------------
+    private Class findClass(String SimpleName) {
+        String full = this.getClass().getName();
+        String dogPath = full.replace(this.getClass().getSimpleName(), SimpleName);
+        try {
+            return Class.forName(dogPath);
+        } catch (ClassNotFoundException e) {
+            fail("\nERROR:Тест не пройден. Класс " + SimpleName + " не найден.");
+        }
+        return null;
+    }
+
+    private Method checkMethod(String className, String methodName, Class<?>... parameters) throws Exception {
+        Class aClass = this.findClass(className);
+        try {
+            methodName = methodName.trim();
+            Method m;
+            if (methodName.startsWith("static")) {
+                methodName = methodName.replace("static", "").trim();
+                m = aClass.getDeclaredMethod(methodName, parameters);
+                if ((m.getModifiers() & Modifier.STATIC) != Modifier.STATIC) {
+                    fail("\nERROR:Метод " + m.getName() + " должен быть статическим");
+                }
+            } else
+                m = aClass.getDeclaredMethod(methodName, parameters);
+            m.setAccessible(true);
+            return m;
+
+        } catch (NoSuchMethodException e) {
+            System.err.println("\nERROR:Не найден метод " + methodName + " либо у него неверная сигнатура");
+            System.err.println("ERROR:Ожидаемый класс: " + className);
+            System.err.println("ERROR:Ожидаемый метод: " + methodName);
+            return null;
+        }
+    }
+
+    private Method findMethod(Class<?> cl, String name, Class... param) {
+        try {
+            return cl.getDeclaredMethod(name, param);
+        } catch (NoSuchMethodException e) {
+            fail("\nERROR:Тест не пройден. Метод " + cl.getName() + "." + name + " не найден\n");
+        }
+        return null;
+    }
+
+    private Object invoke(Method method, Object o, Object... value) {
+        try {
+            method.setAccessible(true);
+            return method.invoke(o, value);
+        } catch (Exception e) {
+            System.out.println(e.toString());
+            fail("\nERROR:Не удалось вызвать метод " + method.getName() + "\n");
+        }
+        return null;
+    }
+
+
+    //метод находит и создает класс для тестирования
+    //по имени вызывающего его метода, testTaskA1 будет работать с TaskA1
+    private static Testing02 run(String in) {
+        return run(in, true);
+    }
+
+    private static Testing02 run(String in, boolean runMain) {
+        Throwable t = new Throwable();
+        StackTraceElement trace[] = t.getStackTrace();
+        StackTraceElement element;
+        int i = 0;
+        do {
+            element = trace[i++];
+        }
+        while (!element.getMethodName().contains("test"));
+
+        String[] path = element.getClassName().split("\\.");
+        String nameTestMethod = element.getMethodName();
+        String clName = nameTestMethod.replace("test", "");
+        clName = clName.replaceFirst(".+__", "");
+        clName = element.getClassName().replace(path[path.length - 1], clName);
+        System.out.println("\n---------------------------------------------");
+        System.out.println("Старт теста для " + clName);
+        if (!in.isEmpty()) System.out.println("input:" + in);
+        System.out.println("---------------------------------------------");
+        return new Testing02(clName, in, runMain);
+    }
+
+    //-------------------------------  тест ----------------------------------------------------------
+    public Testing02() {
+        //Конструктор тестов
+    }
+
+    //переменные теста
+    private String className;
+    private Class<?> aClass;
+    private PrintStream oldOut = System.out; //исходный поток вывода
+    private PrintStream newOut; //поле для перехвата потока вывода
+    private StringWriter strOut = new StringWriter(); //накопитель строки вывода
+
+    //Основной конструктор тестов
+    private Testing02(String className, String in, boolean runMain) {
+        //this.className = className;
+        aClass = null;
+        try {
+            aClass = Class.forName(className);
+            this.className = className;
+
+        } catch (ClassNotFoundException e) {
+            fail("ERROR:Не найден класс " + className + "\n");
+        }
+        InputStream reader = new ByteArrayInputStream(in.getBytes());
+        System.setIn(reader);   //перехват стандартного ввода
+        System.setOut(newOut);  //перехват стандартного вывода
+
+        if (runMain) //если нужно запускать, то запустим, иначе оставим только вывод
+            try {
+                Class[] argTypes = new Class[]{String[].class};
+                Method main = aClass.getDeclaredMethod("main", argTypes);
+                main.invoke(null, (Object) new String[]{});
+                System.setOut(oldOut); //возврат вывода, нужен, только если был запуск
+            } catch (Exception x) {
+                x.printStackTrace();
+            }
+    }
+
+    //проверка вывода
+    private Testing02 is(String str) {
+        assertTrue("ERROR:Ожидается такой вывод:\n<---начало---->\n" + str + "<---конец--->",
+                strOut.toString().equals(str));
+        return this;
+    }
+
+    private Testing02 include(String str) {
+        assertTrue("ERROR:Строка не найдена: " + str + "\n", strOut.toString().contains(str));
+        return this;
+    }
+
+    private Testing02 exclude(String str) {
+        assertTrue("ERROR:Лишние данные в выводе: " + str + "\n", !strOut.toString().contains(str));
+        return this;
+    }
+
+
+    //логический блок перехвата вывода
+    {
+        newOut = new PrintStream(new OutputStream() {
+            private byte bytes[] = new byte[1];
+            private int pos = 0;
+
+            @Override
+            public void write(int b) throws IOException {
+                if (pos == 0 && b == '\r') //пропуск \r (чтобы win mac и linux одинаково работали
+                    return;
+                if (pos == 0) { //определим кодировку https://ru.wikipedia.org/wiki/UTF-8
+                    if ((b & 0b11110000) == 0b11110000) bytes = new byte[4];
+                    else if ((b & 0b11100000) == 0b11100000) bytes = new byte[3];
+                    else if ((b & 0b11000000) == 0b11000000) bytes = new byte[2];
+                    else bytes = new byte[1];
+                }
+                bytes[pos++] = (byte) b;
+                if (pos == bytes.length) { //символ готов
+                    String s = new String(bytes); //соберем весь символ
+                    strOut.append(s); //запомним вывод для теста
+                    oldOut.append(s); //копию в обычный вывод
+                    pos = 0; //готовим новый символ
+                }
+
+            }
+        });
+    }
+}
