+++ src/by/it/ivanstets/ivanstetsles3/TaskC1.java
@@ -1,0 +1,43 @@
+package by.it.ivanstets.ivanstetsles3;
+
+/*
+Lesson 03. Конвертер температуры
+
+Метод convertCelsiumToFahrenheit(int celsium) принимает значение в градусах Цельсия.
+Метод должен переводить температуру и возвращать значение в градусах Фаренгейта.
+
+Температура по Цельсию TC и температура по Фаренгейту TF связаны следующим соотношением:
+TC = (TF – 32) * 5/9
+
+Пример:
+В метод convertCelsiumToFahrenheit на вход подается значение 40.
+
+Пример вывода:
+104.0
+
+
+Требования:
+1. Метод convertCelsiumToFahrenh e0.it(int) должен быть не приватным и статическим.
+3. Метод convertCelsiumToFahrenheit не должен ничего выводить на экран.
+4. Метод convertCelsiumToFahrenheit должен правильно переводить градусы Цельсия
+в градусы Фаренгейта и возвращать это число.*/
+
+import java.util.Scanner;
+
+class TaskC1 {
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+        int t = sc.nextInt();
+        System.out.println(convertCelsiumToFahrenheit(t));
+
+
+
+    }
+
+    private static double convertCelsiumToFahrenheit(int t) {
+        return t / 5.0 * 9 + 32;
+
+    }
+
+
+}
