+++ src/by/it/ivanstets/ivanstetsles4/TaskA3.java
@@ -1,0 +1,43 @@
+package by.it.ivanstets.ivanstetsles4;
+/*
+Существует ли пара?
+Ввести с клавиатуры три целых числа.
+Определить, имеется ли среди них хотя бы одна пара равных между собой чисел.
+Если такая пара существует, вывести на экран числа через пробел.
+Если все три числа равны между собой, то вывести все три.
+
+Требования:
+1. Программа должна считывать числа c клавиатуры.
+2. Программа должна содержать System.out.println() или System.out.print()
+3. Если два числа равны между собой, необходимо вывести числа на экран.
+4. Если все три числа равны между собой, необходимо вывести все три.
+5. Если нет равных чисел, ничего не выводить.
+
+Пример для чисел 1 2 2:
+2 2
+
+Пример для чисел 2 2 2:
+2 2 2
+
+ */
+import java.util.Scanner;
+
+public class TaskA3 {
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+        int a = sc.nextInt();
+        int b = sc.nextInt();
+        int c = sc.nextInt();
+        if (a==b && b==c)
+            System.out.println(a +" " + b + " " + c);
+        else if (b==a)
+            System.out.println(b + " " + a);
+        else if (b==c)
+            System.out.println(b + " " + c);
+        else if (a==c)
+            System.out.println(a + " " + c);
+
+
+    }
+
+}
