+++ src/by/it/ivanstets/ivanstetsles4/TaskB3.java
@@ -1,0 +1,43 @@
+package by.it.ivanstets.ivanstetsles4;
+/*
+С клавиатуры вводится некоторое число.
+Пусть считается сумма int sum=1+2+3+4+5+... и т.д.
+Каждая цифра суммы выводится в консоль через запятую (без пробела).
+Остановите вывод до того, как сумма превысит введенное число.
+
+Например:
+Ввод:
+122
+Вывод:
+1,3,6,10,15,21,28,36,45,55,66,78,91,105,120,
+
+Ввод:
+15
+Вывод:
+1,3,6,10,15,
+
+Ввод:
+14
+Вывод:
+1,3,6,10,
+
+
+ */
+
+import java.util.Scanner;
+
+public class TaskB3 {
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+        int a = sc.nextInt();
+        int sum = 0;
+        int i = 0;
+        while (sum < a) {
+            i++;
+            sum = sum + i;
+            if (sum > a) break;
+            System.out.print(sum + ",");
+        }
+    }
+}
+
